Acoustic Evenness Index - list input

Calculates the Acoustic Evenness Index for all the files in a list, with extended parameter options. It uses parallel processing with all but one of the available cores. Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024.

aei_list(
  audio.list,
  save.csv = TRUE,
  csv.name = "aei_results.csv",
  freq.res = 50,
  w.fun = "hanning",
  min.freq = 0,
  max.freq = 10000,
  n.bands = 10,
  cutoff = 5,
  norm.spec = FALSE,
  noise.red = 2,
  rm.offset = TRUE,
  props = TRUE,
  prop.den = 1,
  n.cores = -1
)

Arguments

audio.list

a list with the audio files to import.

save.csv

logical. Whether to save a csv in the working directory.

csv.name

character vector. When 'save.csv' is TRUE, optionally provide a file name.

w.fun

window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".

min.freq

minimum frequency to compute the spectrogram

max.freq

maximum frequency to compute the spectrogram

n.bands

number of bands to split the spectrogram

cutoff

dB threshold to calculate energy proportions (if norm.spec = FALSE, set to 5 or above)

norm.spec

logical. Whether to normalize the spectrogram (not recommended) or not (normalized spectrograms with different SNR are not comparable).

noise.red

numeric. noise reduction (subtract median from the amplitude values); 1=rows, 2=columns, 3=none.

rm.offset

logical. Whether to remove DC offset before computing aei (recommended) or not.

props

logical. Whether to store the energy proportion values for each frequency band and channel (default) or not.

prop.den

numeric. Indicates how the energy proportion is calculated by manipulating the denominator.

n.cores

The number of cores to use for parallel processing. Use n.cores = -1 to use all but one core. Default is NULL (single-core processing).

frew.res

the frequency resolution (Hz per bin) to use. From this value the window length for the FFT will be calculated (sampling rate / frequency resolution).

Value

a tibble (data frame) with the aei values for each channel (if stereo), metadata, and the parameters used for the calculation.

Details

Options for prop.den: 1 = The original calculation from the "soundecology" package is applied. The denominator of the proportion equals to all the cells in the same frequency band. 2 = The "whole population across species" equals the cells above the decibel threshold across the spectrogram (up to 'max.freq') 3 = The "whole population across species" equals the cells above the decibel threshold across the whole spectrogram (up to the Nyquist frequency. This might return a smaller range of values. It uses parallel processing with all but one of the available cores. Optimized to facilitate working with a list of audio files before importing them into R. The working directory should be set to the folder containing the files. Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024

Examples

files <- list.files(pattern=".wav|.WAV")
aei_list(files[1:5])
#> Evaluating the job...
#> 
#> Error in readBin(con, int, n = 4, size = 1, endian = "little", signed = FALSE) %*%     2^c(0, 8, 16, 24): non-conformable arguments