Merge CSV outputs of several indices calculated with SoundEcology2

Merge CSV outputs of several indices calculated with SoundEcology2

Usage

merge_se2_indices(folder_path = NULL, recursive = TRUE)

Arguments

folder_path

character. The path to the folder containing the csv files with the "results.csv" suffix

recursive

logical. If TRUE, it looks for csv files in the subfolders.

Value

A tibble with the merged tables for different indices.

Examples

if (FALSE) { # \dontrun{
indices_all <- merge_se2_indices("path_to_folder", recursive = TRUE)
} # }