Narrow-Band Activity Index

This index describes the relative amount of narrow-band persistent sound activity, like that of Cicadas and Orthopterans. This index can be used to evaluate insect activity and their influence on other soundscape metrics (e.g., summary acoustic indices).

Usage

nbai(
  wave,
  channel = "left",
  hpf = 250,
  freq.res = 50,
  cutoff = -60,
  activity.cutoff = 10,
  spectrogram = FALSE,
  dark.plot = TRUE,
  plot.title = deparse(substitute(wave)),
  verbose = TRUE
)

Arguments

wave

A Wave object

channel

Character. Options are "left", "right", "mix", or "each". The "mix" option creates a temporary mono file from a mix of both channels before calculating the index. Default is "left".

hpf

Numeric. High-pass filter. The default (500 Hz) should be used always for consistency unless signals of interest are below that threshold.

freq.res

Numeric. Frequency resolution in Hz. This value determines the "height" of each frequency bin and, therefore, the window length to be used (sampling rate / frequency resolution).

cutoff

Numeric. Cutoff threshold defining the sounds that will be analyzed, in dBFS.

activity.cutoff

Numeric. Cutoff percent activity. Only the frequency bands active equal or above this percentage will be considered as "active" in the active band statistics.

spectrogram

Logical. Whether to plot the binary spectrogram used for the analysis. Allowed only when Wave is mono or when one channel is selected from a stereo file.

dark.plot

Logical. If true (default) a the binary spectrogram will have a black background.

verbose

Logical. If TRUE, details of dynamic range will be printed on the console.

Value

A list containing: 1) A binary spectrogram (if mono), 2) tibble with the Narrow-Band Activity Index (NBI) summary statistics, and 3) a tibble with NBI spectral, which number of rows equals the number of frequency bins in the analysis.

Examples

nbai(wave, channel = 'left', plot = TRUE, verbose = TRUE)
#> Error: object 'wave' not found