Normalized Difference Soundscape Index - list imput

Normalized Difference Soundscape Index (NDSI) from REAL and Kasten, et al. 2012. The NDSI seeks to "estimate the level of anthropogenic disturbance on the soundscape by computing the ratio of human-generated (anthrophony) to biological (biophony) acoustic components found in field collected sound samples" (Kasten, et al. 2012). This version is optimized to work with lists of files.

ndsi_list(
  audio.list,
  save.csv = FALSE,
  csv.name = "ndsi_results.csv",
  freq.res = 50,
  anthro.min = 1000,
  anthro.max = 2000,
  bio.min = 2000,
  bio.max = 11000,
  rm.offset = TRUE,
  n.cores = -1
)

Arguments

audio.list

a list of audio files from the working directory.

save.csv

logical. Whether to save a csv in the working directory.

csv.name

character vector. When 'save.csv' is TRUE, optionally provide a file name.

freq.res

numeric. The frequency resolution to use (Hz per bin) which will determine the window length for the FFT (sampling rate / frequency resolution).

anthro.min

minimum value of the range of frequencies of the anthrophony.

anthro.max

maximum value of the range of frequencies of the anthrophony.

bio.min

minimum value of the range of frequencies of the biophony.

bio.max

maximum value of the range of frequencies of the biophony.

rm.offset

logical. Whether to remove the DC offset.

n.cores

The number of cores to use for parallel processing. Use n.cores = -1 to use all but one core. Default is NULL (single-core processing).

Value

a wide format tibble with NDSI values per channel (if stereo), parameters used and audio metadata

Details

It uses parallel processing with all but one of the available cores. Optimized to facilitate working with a list of audio files before importing them in The working directory should be set to the folder containing the files. Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024

Examples

files <- list_waves(path/to/folder)
#> Error in list_waves(path/to/folder): object 'path' not found
ndsi_list(files[1:5])
#> Evaluating the job...
#> 
#> Error in eval_tidy(xs[[j]], mask): object 'files' not found