Check the Integrity of a Wave File Dataset

This function parses all the WAVE files in a folder and checks their size, reporting corrupted files.

wave_integrity(
  folder,
  type = "lines",
  ggplot = TRUE,
  plot.file = "wave_integrity_plot.png",
  log.file = "wave_integrity_log.txt",
  n.cores = -1,
  dump.folder = TRUE,
  tail.folder = TRUE
)

Arguments

folder

Character. The path to the folder containing the wave files.

type

Character. Type of plot to use in the geom call when ggplot is TRUE. Options are "lines", "bars", and "points".

ggplot

Logical. If TRUE, the plot is produced with ggplot2. If FALSE, base plot is used.

plot.file

Character. A name for the output plot. Include the extension in the string.

log.file

Character. A name for the log file (text file) output.

n.cores

Numeric. Number of cores to use in parallel processing. Defults to NULL (no parallel processing).

dump.folder

Logical. If TRUE, a 'dump' folder is created in the working directory and all the DUMP files are moved to that folder.

tail.folder

Logical. If TRUE, a 'tail' folder is created in the working directory and all files belonging to dates after the last date with complete recordings are moved to that folder.

Value

A tibble with summary data on file size, including the last day with complete recordings and days with corrupted files (i.e., smaller than most).

Examples

integrity <- wave_integrity("pathToFolder", ggplot = FALSE)
#> Analyzing the folder... please wait...
#> Error in wave_integrity("pathToFolder", ggplot = FALSE): No .wav files found in the specified folder.