Calculate BBAI for all the Files in a Folder

This function processes an sound signal to detect broadband activity by identifying 'clicks' based on time-frame-wise (i.e., column-wise) amplitude changes in the spectrogram. It computes statistics related to click height, variance, and centroid frequency, and can plot a spectrogram with detected clicks highlighted. The function also classifies whether the signal contains noise or insect based on the variance and centroid frequencies of the clicks.

bbai_folder(
  folder,
  channel = "each",
  hpf = 0,
  rm.offset = TRUE,
  freq.res = 50,
  cutoff = -60,
  click.length = 10,
  difference = 10,
  gap.allowance = 2,
  verbose = FALSE,
  output.csv = "bbai_results.csv",
  n.cores = -1
)

Arguments

folder

Character. The path to a folder with the wave files to analyze.

channel

Character. If Wave is stereo and you want to use only one channel, pass either "left" or "right" to this argument. If you want to analyze a mix of both channels, select "mix". If NULL (default), results are returned for each channel.

hpf

Numeric. High-pass filter. The default (500 Hz) should be used always for consistency unless signals of interest are below that threshold.

rm.offset

Logical. Should the DC offset be removed from the sound signal? Defaults to TRUE.

freq.res

Numeric. Frequency resolution in Hz. This value determines the "height" of each frequency bin and, therefore, the window length to be used (sampling rate / frequency resolution).

cutoff

Numeric. The amplitude threshold (in dBFS) for removing low-amplitude values in the spectrogram. Default is -50.

difference

Numeric. The maximum difference in amplitude between adjacent frequency bins to be considered part of a single 'click'. Default is 20.

gap.allowance

Numeric. The size of gaps (in frequency bins) allowed between contiguous parts of a click. Default is 2. Gaps larger than this value will split clicks.

verbose

Logical. If TRUE, details of dynamic range will be printed on the console.

n.cores

The number of cores to use for parallel processing. Use n.cores = -1 to use all but one core. Default is NULL (single-core processing).

click.height

Numeric. The minimum height (in frequency bins) for a detected click to be kept. Default is 10.

plot

Logical. Should a spectrogram with highlighted clicks be plotted? Default is TRUE.

dark.plot

Logical. Should the plot use a dark theme (black background)? Default is FALSE.

plot.title

Character. The title for the plot, if plot is TRUE. Default is NULL.

Value

A tibble containing the following columns:

index: The name of the index. Useful later when merging data with other indices.

value: The number of clicks detected in the recording.

mean: The mean click height (in frequency bins).

variance: The variance of the click height.

sd: The standard deviation of the click height.

Examples

bbai_folder(path/to/folder)
#> Error: object 'path' not found