Generate a Diel Spectrogram from multiple Audio Recordings

Creates a spectrogram visualization of daily acoustic patterns using multiple recordings, with options for noise floor clipping and parallel processing. Specifically, it takes the variance of acoustic energy in each frequency bin for each recording and merges the resulting vectors into a summary spectrogram.

Usage

diel_spectrogram(
  folder = NULL,
  recursive = TRUE,
  list = NULL,
  title = "",
  freq_res = 50,
  n.cores = -1,
  cutoff = -100,
  max_amp = -10,
  plot = TRUE,
  save_plot = TRUE,
  save_to = "./diel_plots",
  dc_on = NULL,
  dc_off = NULL,
  target_dates = NULL
)

Arguments

folder

Path to the folder containing WAV files (default: current working directory).

recursive

Logical. Whether to search files in subfolders. Default: TRUE.

list

A list with WAV files to parse. The waves must be in the working directory (default: NULL).

title

Character string. Plot title (default: empty).

freq_res

Frequency resolution in Hz for FFT (default: 50)

n.cores

Number of cores for parallel processing (-1 for all available cores)

cutoff

Minimum dBFS value (noise floor). Values below this will be clipped (default: -100)

max_amp

Maximum dBFS value to be represented by the color scale.

plot

Whether to generate the plot (default: TRUE)

save_plot

Logical. Whether to save the plot

save_to

Character. If save_plot = TRUE, provide a path to the folder where the file should be stored. If no path is provided, a new folder "diel_plots" is created in the current working directory (default) to save the exported files.

dc_on

Number of minutes where the recorded was "on". Used to identify expected number of files per day.

dc_off

Number of minutes where the recorded was "off". Used to identify expected number of files per day.

target_dates

Either a single date or multiple dates in the format "YYYY-MM-DD" to be analyzed. If NULL (default), all the dates in the folder will be parsed. To provide a range, see Details.

Value

A list containing:

plot - The ggplot2 spectrogram object

spectral_data - Wide-format spectral data (freq bins × recordings)

plot_data - Long-format data for plotting

file_metadata - File information with extracted metadata

overall_range - The dBFS range of all data

Details

This function parses files in a folder or list and groups them by date, creating a single diel spectrogram for each. To parse a subset range of dates contained in a folder, use seq(). Example: target_dates = seq(as.Date("2022-10-20"), as.Date("2022-10-25), "days")